05.1 - caching wiki, works with graphs/src/requirements.txt
networkx-3.4.2
matplotlib-3.9.2
pyvis-0.3.2
requests
bs4
